+++ android/src/main/java/com/reactnativeandroidwidget/builder/ClickableView.java
@@ -9,12 +9,18 @@ public class ClickableView implements Comparable<ClickableView> {
     private final View view;
     private final String clickAction;
     private final ReadableMap clickActionData;
+    private final String accessibilityLabel;
 
     public ClickableView(String id, View view, String clickAction, ReadableMap clickActionData) {
+        this(id, view, clickAction, clickActionData, null);
+    }
+
+    public ClickableView(String id, View view, String clickAction, ReadableMap clickActionData, String accessibilityLabel) {
         this.id = id;
         this.view = view;
         this.clickAction = clickAction;
         this.clickActionData = clickActionData;
+        this.accessibilityLabel = accessibilityLabel;
     }
 
     public String getId() {
@@ -31,6 +37,10 @@ public class ClickableView implements Comparable<ClickableView> {
 
     public ReadableMap getClickActionData() {
         return clickActionData;
+    }
+
+    public String getAccessibilityLabel() {
+        return accessibilityLabel;
     }
 
     @Override