+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Accessibility (TalkBack) support
+
 ## [0.19.0] - 2026-01-11
 
 ## Added